+++ ICD.Connect.Lighting.Server/LightingProcessorClientDevice.cs
@@ -255,7 +255,7 @@ namespace ICD.Connect.Lighting.Server
 		{
 			base.ClearSettingsFinal();
 
-			m_RpcController.SetPort(null);
+			m_RpcController.SetPort(null, false);
 			ClearCache();
 
 			m_NetworkProperties.ClearNetworkProperties();
@@ -288,7 +288,18 @@ namespace ICD.Connect.Lighting.Server
 				}	
 			}
 
-			m_RpcController.SetPort(port);
+			m_RpcController.SetPort(port, false);
+		}
+
+		/// <summary>
+		/// Override to add actions on StartSettings
+		/// This should be used to start communications with devices and perform initial actions
+		/// </summary>
+		protected override void StartSettingsFinal()
+		{
+			base.StartSettingsFinal();
+
+			m_RpcController.Start();
 		}
 
 		#endregion