+++ ICD.Connect.Lighting.Server/LightingProcessorServer.cs
@@ -129,8 +129,6 @@ namespace ICD.Connect.Lighting.Server
 			Subscribe(m_Server);
 
 			UpdateCachedOnlineStatus();
-
-			InitializeClients();
 		}
 
 		/// <summary>
@@ -1097,7 +1095,6 @@ namespace ICD.Connect.Lighting.Server
 				};
 			}
 
-			tcpServer.Start();
 			SetServer(tcpServer);
 		}
 
@@ -1159,6 +1156,18 @@ namespace ICD.Connect.Lighting.Server
 
 			m_OccupancyControlsByRoom.Clear();
 			m_OccupancyControlForRooms.Clear();
+		}
+
+		/// <summary>
+		/// Override to add actions on StartSettings
+		/// This should be used to start communications with devices and perform initial actions
+		/// </summary>
+		protected override void StartSettingsFinal()
+		{
+			base.StartSettingsFinal();
+
+			m_Server.Start();
+			InitializeClients();
 		}
 
 		private void LogApplySettingsError(string message)